orchestrates/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.15",
    "azure-cognitiveservices-speech>=1.45.0",
    "deepgram-sdk>=4.8.0",
    "fastapi>=0.111.1",
    "google-auth>=2.40.3",
    "google-cloud-storage>=3.4.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "langchain>=0.3.27",
    "loguru>=0.7.3",
    "nltk>=3.9.1",
    "openai>=1.98.0",
    "pinecone>=7.3.0",
    "pyaudio>=0.2.14",
    "pydantic>=2.11.7",
    "pyngrok>=7.2.12",
    "python-dotenv>=1.1.1",
    "redis>=5.3.1",
    "sentry-sdk>=2.35.0",
    "sift-stack-py>=0.9.1",
    "sounddevice>=0.4.7",
    "starlette>=0.37.2",
    "tiktoken>=0.7.0",
    "twilio>=9.7.0",
    "uvicorn>=0.30.6",
    "vocode>=0.1.113",
]
